backend/requirements.txt
# backend/requirements.txt
flask
flask-cors
python-dotenv
cohere
deep_translator
python-docx